Rename folders in ORM, split forgotten class with Django in it.
Two files (calculation/inline.py and calculation/job/__init__.py) were
using Django specific code. This needed to move JobCalculation as an
abstract class with a concrete implementation.

The way it is organized is now as follow:
* Generic class (that don't call directly django's code) should be put
orm/, like before.
* If a class (or function) needs to have some part specific to the
database backend implementation, it should then be split.
    * The generic part should be put into implementation/general, following
the same folder architecture that would be expected in orm/.
    * The concrete implementation would then reside into specific folder
    (such as django), extending the "general" class.
    * The orm/implementation/__init__.py should be edited to import the
    correct implementation depending on the backend. If circular import
    arise, a new file can be created.
    * In orm/, the folder organisation used inside implementation/general
    should be preserved. If a file is in
    implementation/general/foo/bar.py, then a corresponding file
    orm/foo/bar.py should be created,, only to import the class in
    implementation/general.

From the user point of view, they should not import anything that is inside
orm/implementation.