orm/data: implement importstring/importfile
These class methods work similar to the export() method by peeking into
the child classes for `_parser_*` functions and calling them based on
the given format.

One difference is that the `importstring()`is public (does not start
with a `_`) since it may be useful for users running calculations via a
runscript.

Note: this loads files completely into memory which may be not desired
for big files.