TrajectoryData: implement position import from XYZ
Since a XYZ file contains only positions (or velocities, or energies)
and the import* functions do not allow for passing multiple
strings/files or additional attributes, we simply distinguish by
creating a subformat `xyz_pos` instead of simply `xyz`.

Furthermore we have to impose some restrictions on the order of the
import to be able to validate the data passed to the import* functions
before setting the respective arrays.

Usage:
```
  from aiida.orm.data.array.trajectory import TrajectoryData

  t = TrajectoryData()
  # get sites and number of timesteps
  t.set_array('steps', arange(ntimesteps))
  t.set_array('symbols', array([site.kind for site in s.sites]))
  t.importfile('md-test-calculation/AIIDA-PROJECT-pos-1.xyz', 'xyz_pos')
```