Split load_dbenv with an additionnal load_profile method
The splits allows to simply load the profile without the DB
environnement. This is needed for the SqlAlchemy tests, because the
connection needs to be instanciated in a different way than usual.
In particular, we need to access the test profile settings, and then
instanciate the database connection ourself.

The same things has been added for Django; this might not be necessary,
but was here for consistency. It shouldn't change anything besides
adding some more flexibility.