Modify migration to be more iterative, add debug message
Instead of loading everything into memory, and flushing everything at
once, we're only batch loading with size `group_size`, and only the `id`
and `attributes` columns of Node. At the end of each batch, we're
flushing everything and expunging (removing the object from the session)
the object. The memory use should be constant.