Revert to recursive loads_json, but without list allocation
The iterative version was slower than the recursive one. The main reason
in my opinion is list allocation. To be able to iteratively go through
the dictionary, you need to manage a stack (a list in Python) of element
you visit. Because you don't know the size or depth of the dict in
advance, this means at each new "branch" (nested list or dict), you have
to allocate again, and it seems to be heavier to so than to recursively
go trough the dict.