Fix circular import issue with Calculation/JobCalculation
The issue here come from JobCalculation needing specific database access
(using DbCalc), and inheriting from Calculation, which depends on a
concrete implementation of Node. Calculation was thus defined in
orm/calculation/init.py, and JobCalculation in impl/ folder.

To fix this, Calculation has now an abstract version (which is in fact
the whole Calculation class as before) in general, but w/o inheriting
from Node. The concrete Calculation impl. is simply:

class Calculation(AbstractCalculation, Node):
     pass

This way, the JobCalculation import can work correctly. This gymnastic
has to be done to not break the import statement of current AiiDA
program. The current organization could probably be refactor a bit if
one is willing to modify the way some classes are imported.