Abstracted the parsers.py tests for SqlAlchemy and Django
Note that the SqlAlchemy implementation uses the method
import_data_sqla from importexport.py which contains a bug
that causes the tests to fail. Any exception will trigger
the rollback, however a try/except construct is used to
retry the generation of a private key, when a duplicate is
generated where a unique value is required.

On line 1214 the basename of the unique key is generated
based on the current time, but the accuracy is only to within
the second. Since multiple keys are generated within the
same second, duplicates are generated, in which case a
'counter' modifier is added through the triggering of a
UniquenessError exception, see line 1228. Spyros will improve
the prevention of duplicate primary keys