common/utils: fix XYZ parser failure with integers
When parsing XYZ cells where the numbers are pure integers, the parser
sometimes fails to recognize the cell.

For example the following entry:

```
2
Silver dimer; experimental structure form simard01; s
Ag 0 0 0
Ag 0 0 2.5335
```

The reason is, that an OR operator acts on everything inside the
current group. So the regex parser decided between (space, +/-, float) and
(integer or possible float, E-0) which worked rather accidentally.

This can nicely be visualized on https://www.debuggex.com/

The second big issue is the specification for the character classes
`[ ... ]`: inside those, the regex parser will _not_ ignore spaces, even
with the re.X flag, so `[E | e]?` would then literally match 0 or 1 of
E, e, | and ' '.

This fixes issue #293