add verdi-plug next to verdi
verdi-plug is just a click group that loads aiida.cmdline entry points.
It allows separately developping a plugin-enabled version of the verdi
commandline. This will be useful for testing plugins that are in
development against. The mid- to longterm intention is for the verdi-plug
system to replace the current verdi.