Add a hack to JobProcess to circumvent an earlier hack in certain JobCalculations
The MatdynCalculation can take a parent_calculation as an input, however
it is not specified in the accepted use_methods of the inherited
NamelistsCalculation. As a hack, someone has added a literal use_parent_calculation
method in MatdynCalculation, which now has to be called manually.

In the new workflow system this is not possible and so it can not be passed
as an input to the spec, as the spec will only try use methods that are defined
in the calculation's namelist. As a hack, JobProcess will now, if it cannot
find the specified key in its inputs, directly try to call the method and
raise and exception if it does not exist.