Add workaround for RETURN links not existing in old system
With the new FrozenDict data type being able to group the outputs
of a Calculation, to serve as an input for a new Calculation, hence
emulating the possibilites of old workflows, to fetch the corresponding
nodes through the links we should have to look for RETURN links.

However, as the old style workflows never made RETURN links, doing this
would return an empty dictionary for the FrozenDict leading to an
exception. As a hack, we now merge both RETURN and CREATE links