Add a new custom logging level REPORT
The old Workflows, having a separate database table, had a
dedicated column to store informative messages, which are used
heavily by users to follow the flow and progress of the workflow.

The new Workchains, being just a Node, do not have this option
and so logging messages have to be stored elsewhere. We opted to
use the general logging system already in place, which stores
messages in DbLog. However, here the default level was set to
WARNING as we don't want to have the database flooded with
unimportant DEBUG and INFO messages. But since the log messages
for the Workchains would semantically be on INFO level, these
would also be ignored, unless the AiiDA wide log level would
be set to INFO, which would then would log all INFO messages.

To alleviate this we have introduced a new log level

	REPORT = 25

which is reported to the python logging module. To allow this,
the configuration of log levels had to be changed from type
string to type int. Old configurations, with string values for
the log level properties may fail due to this.