Rename LogEntry attribute names to conform with DbLog field names
The DbLog field names were chosen such as to correspond one to
one with the attributes of a 'record' generated by the default
python logging module. This simplifies the storing of such records
directly in the database. Keeping the same names on the Log and
LogEntry class level maintains this simplicity