Verdi quicksetup should set newly created profile to default if not set
The code at the end of verdi quicksetup to set the default profiles
did not work properly if no default profiles were set at all. Which
means that when verdi quicksetup was run for the first time, no default
profile would be set and the user would have to specify it. It makes
more sense for the newly created profile to be setup as the default
profile for all the valid processes