Abstract a layer for the backend and DbLog
Logging to the database was accomplished by adding a
handler to the default python logging module. However,
the handler was only implemented for Django. In order to
be able to define a general handler, an abstract backend
layer was created in aiida.orm.backend which provides an
interface for interacting with the log table, which can
then be implemented for Django and SqlAlchemy.