Abstract python logging module configuration
AiiDA uses the python logging module for all its logging
needs, however the logging was configured only through
Django, by defining the LOGGING config dictionary in
aiida/backends/djsite/settings/settings.py which meant
that when using an SQLA backend, the loggin was never
configured.

The configuration of the loggin module has now been moved
to aiida/__init__.py and the orm interface for both backends
is implemented.

Tests are added that probe both the direct logging interface
of the Log ORM as well as going through the DbHandler that
is attached to the logging system on configuration.

Another test is added in tests/logger.py, that checks that
firing a simple log message will not crash the application
because the DbLogHandler is attached but no backend is
available yet