Add functionality to be able to append values to list in context
In certain situations a user may have a loop in a WorkChain step
in which they launch a process asynchronously and want to simply
add the results to a list in the context, to be used in a later
step.

This was not possible, as one had to use ToContext for this, which
had to be returned as the return value from a step, and so could
effectively only be used once.

As a solution, we expose a method of the WorkChain `insert_intersteps`
which allows to insert one or a list of Intersteps, which will be
executed between the step and its successor. We also implement two
types of Intersteps here: Assign and Append, which take a value and
assign it and append it to some key in the context, respectively.

The key is provided by using a wrapper function ToContext, which takes
the intersteps as key value pairs. The capitalized name is kept to
keep backwards compatibility to when ToContext was an Interstep
subclass.

Note that the saving of instance states of Assign and Append has
not yet been implemented