Ensure WorkChains have a logger that has DbLogHandler set up
The `logger' method of a plum Process will return the default
logger 'plum.process' unless the private member '_logger' has
been explicitly set. For a WorkChain, we want to set the logger
to the logger that is set for its corresponding calculation, as
that will have been correctly setup and will for example have
the DbLogHandler attached. The 'report' method relies on this
handler to log messages to the database from the execution of
the WorkChain steps.

However, in the 'on_create' method of the AiiDA Process class
the method used to check whether a logger had been set was
'logger' which always returns a logger, namely the default of
plum Process if no logger was set and so the logger of the
calculation was never used, breaking the 'report' functionality
of the WorkChain.

The fix is to check directly whether '_logger' is set or not
even though private members should not directly be accessed.