Add reentrant state lock on is_stored property and store method
An issue occurred in the storing of Node's and its Link's in
multithreaded processes, where one thread entered the store method
which after storing the Node within the transaction, set the _to_be_stored
member to False, before the links were stored. A second thread,
operating on the same Node, could now also happen to access the same
Node, query the _to_be_stored value and assume the Node is stored
but the actual transaction of the first thread had not been finished.
Consequently any operations on the Node by the second thread could
fail because the actual node does not yet exist in the database.

To solve this problem, a reentrant state lock was added to the store
method and the is_stored property. All direct reads of _to_be_stored
were changed to go through the is_stored property.