Process WAITING race condition fixed, termination of WaitOnWorkflow polling fixed
The self.interrupt() signal could previously happen inbetween entering the WAITING
state and asking the WaitOn to wait, in which case the flag was reset and the
interrupt was not actioned.  Fixed by adding a clear() method to WaitOn to reset
the internal interrupted flag which is now called before the state lock is released.

WaitOnWorkflow was polling from initialisation (potentially) indefinitely, this caused
the Workflow Emitter thread to remain active even after the Main thread had exited causing
the program to hang.  Now it only polls during the .wait() call (i.e. while another part
of the code is actively waiting for it hence avoiding the thread becoming a zombie).