Add a specific abort function for WorkChain
If a developer wants to abort a Workchain in a certain situation, they
could call the abort method of its Process. However, the timeout for
this method is by default None, which will lead to the thread waiting
on itself to abort indefinitely, as it will only abort at the next state
transition at the earliest. In a Workchain, one will almost always want
to simply abort immediately without timeout. Therefore a new method to
Workchain is added 'abort_nowait', which will call the abort of the
Process with a timeout=0.

The abort method itself is overridden to also add the passed abort
message to the report of the WorkChain