Redesign the UpdateContext Intersteps to make them fully qualified upon construction
Previously, the UpdateContext intersteps Assign and Append, would be constructed
with only the action. The key on which they operated had to be set later through
the set_key method. This is vulnerable to bugs.

These intersteps now require both the key and the action to be passed at construction
time. To facilitate this with the existing user interface of the to_context method of
the WorkChain and the ToContext class, the Assign and Append intersteps have an internal
builder, which will take the action and its build method can then be called with the
key, to construct the complete interstep.