Change custom REPORT loglevel value to avoid clash with stdlib
For the workchains we introduced a new loglevel, REPORT, which
would sit between INFO and WARNING, such that users with the
default WARNING loglevel would experience no difference after
adding the new logging feature for the workchains. We chose the
value 25 originally, which happens to clash with the value for
SUBWARNING which is defined by the multiprocessing module of the
standard python library. Therefore we change the value to 23