De/serialise the contents of WorkChain Context before loading/saving
When a WorkChain is asked to save itself for pickling, it needs to
pickle its Context as well, which will require pickling the internal
content of the Context, which may contain objects, like Data Nodes,
that are not pickleable. Therefore, we need to translate these data
types into a uuid. Upon loading the instance state, the values are
again checked and for example values that resemble a uuid are attempted
to load into corresponding nodes.