Change default behavior of 'verdi work list' and add option
By default the command was listing all the workchain nodes which
would become very slow. Now the default for the '-p/--past-days'
flag is changed from None to 1, so by default only the workchains
of the last day will be listed. The '-a/--all' option is added to
allow for the old behavior of returning all nodes, as it overrides
the '-p/--past-days' flag