Reset the default daemon polling intervals
We leave the workflow step interval at 30 because according to tests
if this is decreased the workflows sometimes tend to crash. The new
workflow engine does not suffer from this problem and can therefore
be reduced to a 5 second polling interval.

For the ssh operations of SUBMIT and RETRIEVE, we can also reduce it
to 10, as these will be usually operated in batches so we do not run
the risk of bombarding many ssh connections