Fix bug in verdi work report and improve legibility
The use of the descendants_of key in the query builder query would
sometimes return wrong workchains as children, simply because they
were related through some shared input node, like e.g. an Int node
for an input. Now we only take direct outputs of.

We also add an option to indent the output based on the level of
the workchain in the call stack