Fix for #536 - passing an integer pk to verdi code show
The verdi code show command was failing when an integer pk was
passed as the AbstractCode get_from_string method was raising
an InputValidationError when a string was passed that could be
converted to an integer.

Since there already exists a general purpose get method Code.get()
that deals properly with the option of a pk or a label, the explicit
get_from_string should only be with strings and proper type should be
ensured by the caller.

The caller in this case is get_code in cmdline/commands/code.py
which now correctly calls get() or get_from_string() based on the
provided input