Fix #538 - virtualenv package missing for clean Ubuntu 16.04 install
For a clean installation of Ubuntu 16.04, the best way of installing
the required virtualenv software is through Ubuntu's own package
manager apt. I have replaced `python-virtualenv` with plain `virtualenv`
in the installation instructions in the documentation. This also made
the installation of this package through pip obsolete and so I removed
it to simply the instructions.

I tested the new instructions on a clean virtual machine with v0.8.1
and everything worked including unit tests. There was one issue during
installation, related to a numpy requirement by psglib. I have added
an entry in the troubleshooting section. A fix is to install numpy
separately through pip first, followed by aiida_core. This shouldn't
be necessary, but it works for now