Backporting new interstep functionality from workchain develop branch
@muhrin and @sphuber have been heavily developing the workchain functionality
in a separate branch. The current state does not allow it to be merged, but
it contains important features that need to be in the 0.9.0 or 0.9.1 release.
One of these features, is the change in how Intersteps work, which allows
the appending of futures to the context. This is an attempt to backport this
functionality from the experimental dev branch to the release candidate.

Note that some additions here are actually in the more recent version of plum
but we want to prevent of having to have another support tag of plum, so
instead we just included those things here. In the future, they will have to
be removed as they belong in plum, e.g. the Savable mixin in aiida/work/intersteps.py