added prompt for profile name to verdi quicksetup
So far, the procedure for choosing a custom profile name
in verdi quicksetup was a bit complicated:
it first checked whether a profile 'quicksetup' was already present
(which already requires sudo access);
if it existed, the user was asked whether to overwrite the profile
or not. by selecting no the user could then choose a new profile name.

This commit adds a prompt for the profile name in verdi quicksetup.
It is the very first prompt and it already has a default value
(i.e. one can simply press "enter" and doesn't lose time).

This is important in particular for setting up test profiles,
as those need to start with 'test_'.

The default name for the database was changed from
    "aiida_qs_<login-name>" to
    "<profile>_<login-name>"
This ensures that for test profiles the database name also starts with
'test_'.