Changed the internal structure of _inputlinks_cache of Node
The old input links cache used a dictionary with the link label
as the key, to store links that were to be created between two
nodes. This clearly put a uniqueness constraint on the link label
regardless of the node and link type, which is clearly undesirable.
This would for example directly disallow two links labeled 'parameters'
one INPUT link and the other a CREATE, even though this should be
perfectly acceptable in the current view and design of nodes and links.

For that reason the structure of the _inputlinks_cache is changed to
use a LinkKey tuple which consists of the source node uuid and the link
label. The values are LinkInfo tuples which contain the full required
information of source node, label and link type.

I removed some stale code, that was not used anywhere: _remove_link_from