Remove references to _remove_link_from from the docs
The Node private method _remove_link_from was removed in the previous
commit, since it had an incorrect function signature and it was not
called anywhere. As an argument it only passed the link 'label' which
would therefore assume a uniqueness constraint on the link labels
independent of source node or link type, which is clearly erroneous.

In this commit, I simply remove the remaining references to this function
from the documentation