Fix bug introduced in aeac53b5b390d0093cad82053e30129c328df483
In the reference commit, the internal structure of the _inputlinks_cache
of the Node class was changed by replacing the key of the dictionary
from being just the link label, to a tuple of the link label and source
node uuid. There used to be a unqiueness constraint therefore just on the
label, which was a problem and the original aim of the fix of the referenced
commit. However, in the _replace_link_from a check for uniqueness on the
link label + link type pair, was kept in place. However, when replacing
this check is not necessary and actually breaks the restart functionality
of the legacy workflows. Therefore, this commit removes that uniqueness
check