Fix exception when calling verdi calculation logshow for WorkCalculation
Since WorkCalculations are a subclass of Calculation nodes, they can also
be passed to 'verdi calculation show', which, if they have log messages
associated with them, will display the message to use 'verdi calculation logshow'
to show these messages. However, this verdi command fetches more than just
log messages but also retrieves attributes unique to JobCalculations, and
so an exception is thrown.

The equivalent verdi command for WorkCalculations is 'verdi work report'.
Ideally 'logshow' should just dispatch the request to the report command
but since the latter is implemented using the click library, but the former
is not, this is not easily accomplishable. Therefore, in the meantime, we
simply print a message that tells the user to use 'verdi work report' instead.