Improve the verdi work list command to resemble calulation list analogue
Unfortunately, the current WorkCalculation nodes do not have a well
defined canonical method of checking their 'state' as for example is
the case with JobCalculations. What it does have are several attribute
keys, FAILED, SEALED and FINISHED that can be set. However, there is
no guarantee for example that if neither FAILED nor FINISHED are not
set, that the calculation is still 'running'.

As a temporary convention we only disply calculations who are set to
be SEALED. The -a/--all-states flag, analogous to the verdi calculation list
command, can be toggled to display all work calculations.