Reverting several commits made for issue #641 regarding input links
The changes made all passed the tests but after more extensive testing
it was found that the changes break the legacy workflows. The changes
allowed the existence of multiple links of the same type with the same
label, as long as the source node was different. The legacy workflow
and calculation plugins bank on the link label for a given link type
to be unique but this was no longer the case. The link system will
need a big overhaul soon anyway, so it will be better to leave the
current state unchanged and do the change in a separate release.

This reverts the following commits:

 * 627a7648e453b34c7865e373237113a5fe532826
 * 11741dd4dc6c95e2a82347be3187ab19b001f9d5
 * 777f80d2e09647447580f13c8af58b201dcd073d
 * aeac53b5b390d0093cad82053e30129c328df483