Fix #653 - Add a tag manually to the append clauses
The original query did not manually specify a tag for the Computer
and User classes that were appended. In recent changes to the
QueryBuilder, in the absence of a specified tag, a unique label
will be automatically generated, which will use a numbered suffix.
However, the verdi calculation cleanworkdir command still anticipated
an unsuffixed key, leading to the KeyError.

To fix this I simply specify the desired tag manually in the append
clause, such that we know exactly which key we should use in the result