Add file locks on pickles when they are running
Processes submitted to the daemon would be persisted between
steps by writing their state to a pickle. Every cycle, the
daemon would detect the pickles in the running directory and
start running them. Since there were no locks being placed
on the pickles, this meant that the same process would be
executed multiple times if the execution time of a given step
would take longer than the daemon cycle.

Here we add a file lock on a process as soon as it is picked
up by the daemon for the duration of its execution. When
trying to start a process from a pickle, the daemon will now
first try to acquire a lock or simply skip it if it fails to
get a lock since it is already in place.