Release locks upon process destroy to prevent deadlock
The locks placed on the pickles were never released even when the
process was destroyed. The files would therefore remain permanently
locked and the daemon could no longer persist them leading to a
complete deadlock.

Adding the on_process_destroyed method to respond to processes
getting destroyed with an appropriate unpersisting of the process
fixes this.