Fix race condition in the workflow ticking engine
Despite adding the file locks on process pickles in the previous
commits, certain steps of processes were still being executed
multiple times when the execution of certain steps exceeded the
ticking interval of the celery task.

The origin lay in the fact that celery tasks are not prevented
by default from overlapping. That is to say, a new celery worker
will be spawned to execute a given task, even if the worker of
the previous cycle has not yet completed its task.

Combined with the way that the tick_workflow_engine loaded the
processes from the pickles into memory, this lead to multiple
workers loading up the same process in the same state and therefore
running the same step multiple times.

The real underlying problem was the delay between the processes being
created from the pickles and the file locks being placed on the pickles.
The ticking engine would first load up all the processes and only
then would loop over them, locking the pickle and running the process.
In between the process being loaded and its pickle being locked, any
other process was free to load it up in preparation of execution.

The solution is to move the creation of pickle lock to the moment
that anyone wants to recreate a process from it in order to run it.