Remove public method to load checkpoint without file lock
The method load_checkpoint allowed anyone to load a checkpoint
from a pickle without placing a lock on the pickle. If the
checkpoint was used to recreate the process and run it, the
same process steps may get executed multiple times. To reduce
the chance of this happening we change the method to be private.
The verdi work checkpoint command now uses a call that will
return a string of the state, such that it cannot be used to
recreate the corresponding process.