Allow namespacing in inputs and outputs of a ProcessSpec
This change was implemented to accommodate the new functionality that
allows a Process to expose the inputs of another Process. The goal
for WorkChain design should be to write well defined abstract components
that can then be composed together by wrapping workchains. To prevent
to having to manually copy over the input specification of a sub workchain
we provide the 'expose_inputs' method that will do this automatically.

To allow for inputs to be separated, we change the inputs of the ProcessSpec
to be a PortNamespace, which allows for arbitrary nesting of Ports.