Implement the 'include' option for ProcessSpec.expose_inputs
A common use case is for a wrapping Process to run multiple instances
of another sub Process and therefore wants to expose its inputs.
However, the required inputs for the sub Process instances may differ
for some inputs, while others can be shared. To accomplish this one
can expose the inputs twice in separate namespaces and expose the
shared inputs on the top level namespace. To simplify the specification
of which inputs to expose where, we introduce the 'include' parameter
for the ProcessSpec.expose_inputs method.

To simply the gathering of inputs from the parsed inputs for a given
sub process over multiple namespaces, we adapt the behaviour of the
exposed_inputs methods to traverse down the namespace tree, collecting
all inputs that are common to the process in question. This agglomerating
behavior can be prevented by passing the argument agglomerate=False