Make TCOD exporter tests conditional on NWChem plugin being installed
Several TCOD exporter tests depend directly on the NWChem pymatgen
parser that is now migrated to a separate plugin repository aiiga-nwchem.
As a result, if the plugin is not installed, the tests will fail. We
introduce a conditional that will test whether the plugin is present
otherwise the relevant tests are skipped