Replace outdated existing_plugins with all_plugins of pluginloader
The old plugin loader defined the function existing_plugins to retrieve
all the plugins for a given category. With the advent of the new
plugin loader, this function was mirrored and relayed back to the old
version and when failing called the list of plugins registered through
the new entry point system. However, the all_plugins function tried
to accomplish the same goal. In this commit we remove the existing_plugins
function from the new plugin loader and replace all calls in the code
with a call to the all_plugins function instead.