Allow for the possibility of entry points to be overriden
There is a use case where a registered entry point needs to be
overriden by a different class from another plugin repository.
The plugin loader needs to be able to check whether a specific
entry point can be overriden by another one if more than one are
registered.

Note, however, that both entry points need to agree that the one
overrides the other in order to prevent accidental overriding.