Fixes #722 issue affecting failed workflows
We overwrote the constructor for Persistence but did not call the super
class.  This was a problem because the super constructor was actually
adding the class as a listener of the MONITOR.  It is the MONITOR that
would tell the persistence when a process had crashed (because of
exception).  This means that the pickle was not being moved to failed
and therefore the workflow would be re-ran.