Write serializer and deserializer for mappings containing Node instances
For task payloads to be communicated with RMQ they need to be serialized
to a string. In the case of AiiDA processes, the inputs will be instances
of the Node class which will have to be serialized to their UUID.
This functionality is needed in many places and should be abstracted to
another place for easy reuse.