Change ProcessLaunchPublisher to accept Process Bundles
The way AiiDA nodes are designed, they have to be stored before
they get a persistent pk, allowing them to be serialized and
deserialized by another process. This is required to allow process
instances to be communicated over RMQ. As a solution we now demand
that a process and its inputs are stored and that a bundle is passed
as an argument to the launch command of the ProcessLaunchPublisher
as opposed to passing the process class and its inputs without
explicitly instantiating it. The ProcessLaunchSubscriber will then
recreate the process instance from the bundle and continue its
execution