Restore part of commit that got removed in latest develop merge
During the last merge of develop into the workflows branch part
of recently added code got accidentally deleted. The definition
of the Transport classed was moved from aiida.transport.__init__
to aiida.transport.transport in the workflows branch, but in doing
so changes introduced in develop in the original file were missed
The relevant commit is:

    05f8372b7edba97b01a6820fd5e3ac4d532d8f24

I have reintroduced the missing snippet of code and the test are
now back to a working state.