Migrate existing export archives for unit tests
The import export functionality has been updated to include
link types in the export and properly restore them when
importing from the archive. A couple of unit tests use a
precreated export archive for testing purposes. These had to
be migrated to the new schema to include the link types.
The script with which this was performed has been included in
the docs tree. We might want to make this functionality
accessible through a 'verdi export migrate' entry point.