During export from parent nodes only follow CREATE links
The current logic of the export functionality is based on
the old link type conception, i.e. there are only INPUT
and CREATE links. With the addition of WorkCalculation nodes
the new link types CALL and RETURN were added. The old export
logic would follow any link types from the included parent
nodes, which now also included WorkCalculation nodes connected
through CALL links. The problem is that the input nodes of those
WorkCalculations would however not be included, but their links
would be. This inconsistency leads to broken export archives.

The temporary fix is to make the export functionality compatible
with the old link conception, which means to only follow CREATE
links from the detected parent nodes.