_append_to_attr and simplified clean_value
It turns out that 'defer'ing the execution of `clean_value` to storing
values in the database actually isn't really useful:
When updating an existing attribute, `clean_value` would anyhow be
executed when calling `get_attr`. Advantage only remains, when you want
to overwrite attributes without reading the existing values.

We don't consider this to be a particularly relevant use case and thus
reduce the options for clean_value to "now" or "never".
This also has the advantage of sparing us the need to store a list of
dirty attributes.