Migrate the export archives used for testing from 0.2 to 0.3
The archives had already been migrated in a previous commit from
0.2 to 0.3, where the link types were added to the archive. Since
then the export format was further adapted and we wanted to include
it in the same version migration. Here we apply the actual migration
to the existing archives used for testing, were the mapping of entity
names is done to make them truely backend agnostig.