Change the state sequence for Process abort
For WorkCalculations to be abortable through the kill() method
the abort sequence of the Process had to be changed a bit. Upon
abort, the ABORTED key will be set on the calculation node and
the process will be put in the STOPPED state through .stop()
The on_destroy method will check the state of the calculation
and seal it when appropriate. The other point where the node
can be sealed is in the ProcessFinaliser which will be hit if
the process fails during a tick