Remove the backend dependent code from verdi daemon stop
The verdi daemon stop command was backend dependent because
it contained code that tried to reset the timestamps of the
legacy workflow stepper. However, this code is also code in
the verdi daemon start command and as such is not necessary.
Removing this code from verdi daemon stop means that it can
now be called without having to load the database environment.
This solves the problem that occurs when people upgrade their
AiiDA installation and a database migration is needed, for
which they needed to stop their daemon, but the necessity of
having to load the database environment, triggered the very
same database schema check, resulting in an exception.

With this change, people can call verdi daemon stop to stop
the daemon before performing a database migration, even when
there is a database version mismatch