Improve the STATE column for verdi work report
Recently another terminal state was added for WorkCalculations to
be able to support aborting running WorkCalculations. With that
the state picture of a WorkCalculation also became a little bit
better defined. A WorkCalculation has three terminal states:

 * FINISHED
 * FAILED
 * ABORTED

These states are set by the corresponding attribute to the node
and are mutually exclusive. If the WorkCalculation is in neither
one of these three terminal states, it can be considered by
inference to be RUNNING.

A node can also be sealed or unsealed. It can only be sealed until
after it has reached one of the three terminal states. If a node
is sealed but is not in a terminal state, the state is inconsistent
and so we print Unknown. This is to prevent from WorkCalculations
during which the code excepted will be incorrectly labeled
as RUNNING, which would confuse the user.