Implement the concept of retrieve_temporary_list in CalcInfo
For some calculation one wants to retrieve certain files for parsing
but does not want to store them permanently in the repository because
for example they may take up a lot of space. To facilitate for this
use case we introduce the concept of the retrieve_temporary_list member
in the CalcInfo datastructure. Implementations of a Calculation can
opt to provide a list of directives, identical to that of the original
retrieve_list member, that will instruct the exectution manager to
retrieve those files to a local folder for parsing. The difference is
that the files listed in the retrieve_temporary_list member will be
stored in a FolderData node that will not be stored in the database
and which will only be alive for the duration of the parsing. After
parsing is completed the retrieved files are garbage collected.